New API
We can generate all we need from the command dicts, do not repeat the
work. Split into protocol and transport.
Clean up Serial Port and Telnet, ignore TCP for now, since it seem to be
significantly different.

This adds a test for C 356BE, so that it doesn't break again.